+++ lib/twilio/connection.ts
@@ -95,6 +95,12 @@ class Connection extends EventEmitter {
   static toString = () => '[Twilio.Connection class]';
 
   /**
+   * Describes the phone number that is initiating the call.
+   * Null if this call is outbound from this Client, or if the caller is not from PSTN.
+   */
+  readonly callerInfo: Connection.CallerInfo | null;
+
+  /**
    * The custom parameters sent to (outgoing) or received by (incoming) the TwiML app.
    */
   readonly customParameters: Map<string, string>;
@@ -260,6 +266,15 @@ class Connection extends EventEmitter {
     }
 
     this._direction = this.parameters.CallSid ? Connection.CallDirection.Incoming : Connection.CallDirection.Outgoing;
+
+    if (this._direction === Connection.CallDirection.Incoming && this.parameters) {
+      const isFromPSTN: boolean = /^\+?[\d-\(\) ]+$/.test(this.parameters.From);
+      this.callerInfo = isFromPSTN
+        ? { isVerified: this.parameters.StirStatus === 'TN-Validation-Passed-A' }
+        : null;
+    } else {
+      this.callerInfo = null;
+    }
 
     this._mediaReconnectBackoff = Backoff.exponential(BACKOFF_CONFIG);
     this._mediaReconnectBackoff.on('ready', () => this.mediaStream.iceRestart());
@@ -1519,6 +1534,20 @@ namespace Connection {
      * Twilio Voice related error
      */
     twilioError?: TwilioError;
+  }
+
+  /**
+   * Describes the phone number that initiated an incoming call.
+   */
+  export interface CallerInfo {
+    /**
+     * Whether or not Twilio was able to verify whether the caller is authorized to use the number
+     * that this call is from. If true, it is safe to trust that the caller is who they claim to be.
+     * If this is false, it does not mean that the call is faked, only that we were not able to
+     * verify authenticity. Most legitimate calls at the time of implementation will not be
+     * verifiable.
+     */
+    isVerified: boolean;
   }
 
   /**